src/requirements.txt
python-telegram-bot==12.6.1
sentry-sdk==0.14.3
envparse==0.2.0
falcon==2.0.0
celery==4.4.2
flower==0.9.4
redis==3.4.1
flake8==3.7.9
flake8-bugbear==20.1.4
flake8-commas==2.0.0
flake8-isort==3.0.0
flake8-print==3.1.4